# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RobotOps Sentinel API",
    "summary": "Closed-loop robotics supervision backend.",
    "license": {
      "name": "Proprietary hackathon prototype",
      "identifier": "LicenseRef-Proprietary"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development backend"
    }
  ],
  "paths": {
...
    "/api/demo-replay/export/encord": {
      "post": {
        "tags": [
          "demo integrations"
        ],
        "summary": "Export Demo Replay To Encord",
        "description": "Upload before/after incident evidence to Encord when configured.",
        "operationId": "export_demo_replay_to_encord_api_demo_replay_export_encord_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncordExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncordIncidentExportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo-replay/incident/open": {
      "post": {
        "tags": [
          "demo integrations"
        ],
        "summary": "Open Live Incident",
        "description": "Store the before frame for the current live incident.",
        "operationId": "open_live_incident_api_demo_replay_incident_open_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiveIncidentSignalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveIncidentSignalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo-replay/incident/resolve": {
      "post": {
        "tags": [
          "demo integrations"
        ],
        "summary": "Resolve Live Incident",
        "description": "Resolve the live incident and export the stored frames to Encord.",
        "operationId": "resolve_live_incident_api_demo_replay_incident_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiveIncidentSignalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveIncidentSignalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EncordExportRequest": {
        "properties": {
          "before_image_data_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Image Data Url"
          },
          "after_image_data_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Image Data Url"
          },
          "before_image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Image Path"
          },
          "after_image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Image Path"
          },
          "include_openai_report": {
            "type": "boolean",
            "title": "Include Openai Report",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EncordExportRequest",
        "description": "Optional live Encord upload request for demo evidence."
      },
...
      "EncordIncidentExportPacket": {
        "properties": {
          "ontology": {
            "$ref": "#/components/schemas/EncordOntologySpec"
          },
          "incident": {
            "$ref": "#/components/schemas/SafetyIncident"
          },
          "review_sample": {
            "$ref": "#/components/schemas/ReviewSample"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/EncordIncidentFrame"
            },
            "type": "array",
            "minItems": 2,
            "title": "Frames"
          },
          "outcome_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentOutcomeReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ontology",
          "incident",
          "review_sample",
          "frames",
          "metadata"
        ],
        "title": "EncordIncidentExportPacket",
        "description": "Complete Encord-ready incident packet."
      },
      "EncordIncidentExportResult": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ExportStatus"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "provider_sample_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Provider Sample Ids"
          },
          "packet": {
            "$ref": "#/components/schemas/EncordIncidentExportPacket"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "detail",
          "packet"
        ],
        "title": "EncordIncidentExportResult",
        "description": "Result of attempting to upload incident evidence to Encord."
      },
      "EncordIncidentFrame": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "before",
              "after"
            ],
            "title": "Role"
          },
          "evidence_frame": {
            "$ref": "#/components/schemas/EvidenceFrame"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/EncordObjectLabel"
            },
            "type": "array",
            "title": "Objects"
          },
          "classifications": {
            "$ref": "#/components/schemas/EncordFrameClassification"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "evidence_frame",
          "objects",
          "classifications"
        ],
        "title": "EncordIncidentFrame",
        "description": "One before/after frame prepared for Encord incident logging."
      },
...
      "EncordOntologyClassification": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "radio",
              "text"
            ],
            "title": "Kind"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "kind",
          "description"
        ],
        "title": "EncordOntologyClassification",
        "description": "Classification type that should exist in the Encord ontology."
      },
...
      "EncordOntologyObject": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "shape": {
            "type": "string",
            "enum": [
              "bounding_box",
              "polygon"
            ],
            "title": "Shape"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "shape",
          "description"
        ],
        "title": "EncordOntologyObject",
        "description": "Object type that should exist in the Encord ontology."
      },
      "EncordOntologySpec": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/EncordOntologyObject"
            },
            "type": "array",
            "title": "Objects"
          },
          "classifications": {
            "items": {
              "$ref": "#/components/schemas/EncordOntologyClassification"
            },
            "type": "array",
            "title": "Classifications"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "description",
          "objects",
          "classifications"
        ],
        "title": "EncordOntologySpec",
        "description": "Source-of-truth ontology expected by the incident export pipeline."
      },
...
      "ExportStatus": {
        "type": "string",
        "enum": [
          "local_only",
          "export_pending",
          "exported",
          "export_unavailable"
        ],
        "title": "ExportStatus",
        "description": "Review/export availability states."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentOutcomeReport": {
        "properties": {
          "summary": {
            "type": "string",
            "minLength": 1,
            "title": "Summary"
          },
          "worked": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Worked"
          },
          "failed_or_risky": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed Or Risky"
          },
          "recommended_dataset_labels": {
            "$ref": "#/components/schemas/IncidentOutcomeDatasetLabels"
          },
          "operator_review_needed": {
            "type": "boolean",
            "title": "Operator Review Needed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "worked",
          "failed_or_risky",
          "recommended_dataset_labels",
          "operator_review_needed"
        ],
        "title": "IncidentOutcomeReport",
        "description": "Structured analysis of an incident outcome."
      },
...
      "LiveIncidentSignalRequest": {
        "properties": {
          "image_data_url": {
            "type": "string",
            "minLength": 1,
            "title": "Image Data Url"
          },
          "blocking_object": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocking Object"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "image_data_url"
        ],
        "title": "LiveIncidentSignalRequest",
        "description": "Live frame payload sent by the browser for the active incident."
      },
      "LiveIncidentSignalResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "exported",
              "failed"
            ],
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "packet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EncordIncidentExportPacket"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_sample_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Provider Sample Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "detail"
        ],
        "title": "LiveIncidentSignalResponse",
        "description": "Status for a live incident signal."
      },
...
      "ReviewSample": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "before_frame_id": {
            "type": "string",
            "title": "Before Frame Id"
          },
          "after_frame_id": {
            "type": "string",
            "title": "After Frame Id"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "human_decision": {
            "$ref": "#/components/schemas/ReviewDecision"
          },
          "export_status": {
            "$ref": "#/components/schemas/ExportStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "incident_id",
          "before_frame_id",
          "after_frame_id",
          "labels",
          "human_decision",
          "export_status"
        ],
        "title": "ReviewSample",
        "description": "Review/eval sample produced from a closed incident."
      },
      "SafetyIncident": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "state": {
            "$ref": "#/components/schemas/IncidentState"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "before_evidence_frame_id": {
            "type": "string",
            "title": "Before Evidence Frame Id"
          },
          "incident_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "alert_events": {
            "items": {
              "$ref": "#/components/schemas/AlertEvent"
            },
            "type": "array",
            "title": "Alert Events"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClearanceVerification"
              },
              {
                "type": "null"
              }
            ]
          },
          "audit_packet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuditPacket"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_sample": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewSample"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "zone_id",
          "state",
          "severity",
          "before_evidence_frame_id"
        ],
        "title": "SafetyIncident",
        "description": "Operational record for a critical-zone safety violation."
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity",
        "description": "Incident severity levels."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}